Moved enums from pojos to models
