+++ assure/src/main/java/org/learning/assure/model/enums/UserType.java
@@ -1,4 +1,4 @@
-package org.learning.assure.pojo;
+package org.learning.assure.model.enums;
 
 public enum UserType {
     CLIENT, CUSTOMER
