+++ assure/src/main/java/org/learning/assure/model/form/ChannelForm.java
@@ -1,11 +1,10 @@
 package org.learning.assure.model.form;
 
 import lombok.Data;
-import org.learning.assure.pojo.InvoiceType;
+import org.learning.assure.model.enums.InvoiceType;
 
 @Data
 public class ChannelForm {
     private String name;
-
     private InvoiceType invoiceType;
 }
