+++ assure/src/main/java/org/learning/assure/model/form/UserForm.java
@@ -1,7 +1,7 @@
 package org.learning.assure.model.form;
 
 import lombok.Getter;
-import org.learning.assure.pojo.UserType;
+import org.learning.assure.model.enums.UserType;
 
 @Getter
 public class UserForm {
